pyproject.toml
[tool.poetry]
name = "user-expense"
version = "0.1.0"
description = ""
authors = ["bbyte <i1.eng.dev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<4"
Django = "^5.1.3"
djangorestframework = "^3.15.2"
drf-yasg = "^1.21.8"
ruff = "^0.8.0"
django-environ = "^0.11.2"


[tool.ruff]
builtins = ["_"]
exclude = ["test_venv", "venv_test", "venv", ".venv"]
line-length = 120
show-fixes = true
src = ["expenses", "user_expense"]

[tool.ruff.format]
quote-style = "double"

[tool.ruff.lint]
select = ["E", "F", "C", "I", "Q"]

[tool.ruff.lint.mccabe]
max-complexity = 15

[tool.ruff.lint.flake8-quotes]
docstring-quotes = "double"
inline-quotes = "double"

[tool.ruff.lint.pep8-naming]
classmethod-decorators = [
    "pydantic.validator",
    "declared_attr",
    "expression",
    "comparator",
]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


